requirements.txt
django>3.2,<3.3
djangorestframework
markdown
django-filter
python-decouple
djoser
djangorestframework-simplejwt[crypto]
#django-rest-swagger
drf-yasg[validation]
pytest
pytest-django
